+++ src/prelude.rs
@@ -3,6 +3,7 @@ use std::{path::PathBuf, sync::Mutex};
 use once_cell::sync::Lazy;
 
 pub use crate::path::StrictPath;
+use crate::resource::manifest::Os;
 
 pub static VERSION: Lazy<&'static str> =
     Lazy::new(|| option_env!("LUDUSAVI_VERSION").unwrap_or(env!("CARGO_PKG_VERSION")));
@@ -10,16 +11,13 @@ pub static VARIANT: Option<&'static str> = option_env!("LUDUSAVI_VARIANT");
 
 pub type AnyError = Box<dyn std::error::Error>;
 
-pub const WINDOWS: bool = cfg!(target_os = "windows");
-pub const MAC: bool = cfg!(target_os = "macos");
-pub const LINUX: bool = cfg!(target_os = "linux");
-pub const CASE_INSENSITIVE_OS: bool = WINDOWS || MAC;
 pub const SKIP: &str = "<skip>";
 pub const APP_DIR_NAME: &str = "ludusavi";
 const PORTABLE_FLAG_FILE_NAME: &str = "ludusavi.portable";
 pub const INVALID_FILE_CHARS: &[char] = &['\\', '/', ':', '*', '?', '"', '<', '>', '|', '\0'];
 
-pub static STEAM_DECK: Lazy<bool> = Lazy::new(|| LINUX && StrictPath::new("/home/deck".to_string()).exists());
+pub static STEAM_DECK: Lazy<bool> =
+    Lazy::new(|| Os::HOST == Os::Linux && StrictPath::new("/home/deck".to_string()).exists());
 
 // NOTE.2022-11-04 not very pretty singleton like global variable
 pub static CONFIG_DIR: Mutex<Option<PathBuf>> = Mutex::new(None);