+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Added:
   * CLI: Backup comments are now included in the output of the `backups` command.
 * Fixed:
+  * When using Heroic on Linux to run Windows games,
+    save paths in the game install folders are now checked case-insensitively.
+* Fixed:
   * GUI: On the other screen, backup exclusions could be formatted incorrectly
     if you tried to undo/redo before making any changes to them.
   * GUI: On Mac, if a backup included multiple direct children of the root directory,